+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsDataFrameAggregateSuite.scala
@@ -19,38 +19,12 @@
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
 
-import org.apache.spark.sql.{DataFrameAggregateSuite, Row}
-import org.apache.spark.sql.functions._
+import org.apache.spark.sql.DataFrameAggregateSuite
 import org.apache.spark.sql.rapids.utils.RapidsSQLTestsTrait
 
 class RapidsDataFrameAggregateSuite extends DataFrameAggregateSuite with RapidsSQLTestsTrait {
   // example to show how to replace the logic of an excluded test case in Vanilla Spark
   testRapids("collect functions" ) {  // "collect functions" was excluded at RapidsTestSettings
     // println("...")
-  }
-  import testImplicits._
-
-  test("groupBy22") {
-
-    com.nvidia.spark.rapids.profiling.DumpedExecReplayer.main(
-      Array("/tmp/lore/plan_id=290/partition_id=0"))
-    println("xxx")
-    scala.io.StdIn.readLine()
-
-
-    testData2.createOrReplaceTempView("testData2")
-    spark.sql("select a, sum(b) from testData2 where a < (select max(y) from values" +
-      " (1,2),(2,3),(3,4) as data(x,y)    ) group by a").show()
-    println("...")
-    scala.io.StdIn.readLine()
-
-
-    checkAnswer(
-      testData2.groupBy("a").agg(sum($"b")),
-      Seq(Row(1, 3), Row(2, 3), Row(3, 3))
-    )
-    println("...")
-    scala.io.StdIn.readLine()
-
   }
 }
